Add xml to show article thumbnails
